remove assertions
